undo changes for extractFileTitle test
